+++ imports/api/games/server/hooks.js
@@ -4,7 +4,7 @@ import { Games } from '../games';
 import GameUpdater from './updater';
 import SeasonLeagueTeamUpdater from '../../season_league_teams/server/updater';
 import PoolTeamWeeksUpdater from '../../pool_team_weeks/server/updater';
-import PoolTeamRecordsUpdater from '../../pool_team_records/server/updater';
+import PoolTeamHeadToHeadRecordsUpdater from '../../pool_team_head_to_head_records/server/updater';
 
 // performance improvement - https://github.com/matb33/meteor-collection-hooks#afterupdateuserid-doc-fieldnames-modifier-options
 Games.hookOptions.after.update = { fetchPrevious: false };
@@ -20,8 +20,8 @@ Games.after.update((userId, doc, fieldNames, modifier, options) => {
     SeasonLeagueTeamUpdater.updateTeamStats(doc.leagueId, doc.seasonId, doc.homeTeamId);
     SeasonLeagueTeamUpdater.updateTeamStats(doc.leagueId, doc.seasonId, doc.awayTeamId);
 
-    PoolTeamRecordsUpdater.updatePoolTeamByLeagueTeam(doc.leagueId, doc.seasonId, doc.homeTeamId);
-    PoolTeamRecordsUpdater.updatePoolTeamByLeagueTeam(doc.leagueId, doc.seasonId, doc.awayTeamId);
+    PoolTeamHeadToHeadRecordsUpdater.updatePoolTeamByLeagueTeam(doc.leagueId, doc.seasonId, doc.homeTeamId);
+    PoolTeamHeadToHeadRecordsUpdater.updatePoolTeamByLeagueTeam(doc.leagueId, doc.seasonId, doc.awayTeamId);
 
     if (doc.week) {
       PoolTeamWeeksUpdater.updateLeagueTeam(doc.leagueId, doc.seasonId, doc.homeTeamId, doc.week);
