+++ imports/api/pool_team_picks/server/hooks.js
@@ -2,7 +2,7 @@ import { PoolTeamPicks } from '../pool_team_picks';
 import { Pools } from '../../pools/pools';
 import PoolTeamUpdater from '../../pool_teams/server/updater';
 import RatingCalculator from '../../pool_game_interest_ratings/server/calculator';
-import PoolTeamRecordsUpdater from '../../pool_team_records/server/updater';
+import PoolTeamRecordsUpdater from '../../pool_team_head_to_head_records/server/updater';
 
 // performance improvement - https://github.com/matb33/meteor-collection-hooks#afterupdateuserid-doc-fieldnames-modifier-options
 PoolTeamPicks.hookOptions.after.update = { fetchPrevious: false };
