+++ imports/startup/server/register-api.js
@@ -11,7 +11,7 @@ import '../../api/pool_game_interest_ratings/server/publications';
 import '../../api/pool_team_picks/server/hooks';
 import '../../api/pool_team_picks/server/publications';
 
-import '../../api/pool_team_records/server/publications';
+import '../../api/pool_team_head_to_head_records/server/publications';
 
 import '../../api/pool_team_weeks/server/publications';
