+++ imports/ui/components/pools-head-to-head-records.js
@@ -1,17 +1,16 @@
 import { Meteor } from 'meteor/meteor';
 import { Template } from 'meteor/templating';
 import SimpleSchema from 'simpl-schema';
-import { _ } from 'lodash';
 
 import { PoolTeams } from '../../api/pool_teams/pool_teams';
-import { PoolTeamRecords } from '../../api/pool_team_records/pool_team_records';
+import { PoolTeamHeadToHeadRecords } from '../../api/pool_team_head_to_head_records/pool_team_head_to_head_records';
 
 import './pools-head-to-head-records.html';
 
 Template.Pools_head_to_head_records.helpers({
-  poolTeamRecords: () => {
+  poolTeamHeadToHeadRecords: () => {
     const { poolId, seasonId } = Template.currentData();
-    return PoolTeamRecords.find({ seasonId, poolId }, {
+    return PoolTeamHeadToHeadRecords.find({ seasonId, poolId }, {
       sort: { wins: 1 },
       limit: 50,
     });
@@ -29,6 +28,6 @@ Template.Pools_head_to_head_records.onCreated(function () {
   this.getPoolTeam = (poolTeamId) => PoolTeams.findOne(poolTeamId);
 
   this.autorun(() => {
-    this.subscribe('poolTeamRecords.ofPool', this.data.poolId, this.data.seasonId);
+    this.subscribe('poolTeamHeadToHeadRecords.ofPool', this.data.poolId, this.data.seasonId);
   });
 });
